refactor: import errText directly in request handlers
